[bsp][pico] Update SDK to 1.5.1

+++ bsp/raspberry-pico/libraries/pico-sdk/src/rp2_common/hardware_vreg/vreg.c
@@ -8,5 +8,5 @@
 #include "hardware/vreg.h"
 
 void vreg_set_voltage(enum vreg_voltage voltage) {
-    hw_write_masked(&vreg_and_chip_reset_hw->vreg, voltage << VREG_AND_CHIP_RESET_VREG_VSEL_LSB, VREG_AND_CHIP_RESET_VREG_VSEL_BITS);
+    hw_write_masked(&vreg_and_chip_reset_hw->vreg, ((uint)voltage) << VREG_AND_CHIP_RESET_VREG_VSEL_LSB, VREG_AND_CHIP_RESET_VREG_VSEL_BITS);
 }
